sources/ricable/ottomator-agents/foundational-rag-agent/requirements.txt
pydantic-ai>=0.1.12
supabase>=2.0.0
openai>=1.0.0
PyPDF2>=3.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
numpy>=1.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
